+++ source/app/mocks/api/github/rest/username.mjs
@@ -1,0 +1,18 @@
+/** Mocked data */
+  export default function({faker}, target, that, [{username}]) {
+    console.debug(`metrics/compute/mocks > mocking rest api result > rest.repos.getByUsername`)
+    return ({
+      status:200,
+      url:`'https://api.github.com/users/${username}/`,
+      headers: {
+        server:"GitHub.com",
+        status:"200 OK",
+        "x-oauth-scopes":"repo",
+      },
+      data:{
+        login:faker.internet.userName(),
+        avatar_url:null,
+        contributions:faker.random.number(1000),
+      }
+    })
+  }
